🚨 Fix compiler warning C4267 'argument': conversion from 'size_t' to 'const int', possible loss of data

+++ Sandbox/src/Examples/03_SocketNetworking.cpp
@@ -34,7 +34,7 @@ int doExample03() {
 
         // Send message to the server
         const char* message = "Hello from client.";
-        socket.send(message, strlen(message));
+        socket.send(message, (int)strlen(message));
 
         // Receive message from the server
         char buffer[256] = {};
@@ -68,7 +68,7 @@ int doExample03() {
 
         // Send message to client
         const char* message = "Hello from server.";
-        client->send(message, strlen(message));
+        client->send(message, (int)strlen(message));
 
     }
